digitizer and calibrator check output node before creating, fixes for virtual functions

+++ simulation/g4simulation/g4cemc/RawTower.h
@@ -76,7 +76,7 @@ class RawTower : public PHObject {
   virtual void clear_g4showers() {}
 
   virtual int get_tower_type() const { PHOOL_VIRTUAL_WARN("get_tower_type()"); return -1; }
-  virtual void set_tower_type (const double ) { PHOOL_VIRTUAL_WARN("set_tower_type()"); return  ; }
+  virtual void set_tower_type (const int ) { PHOOL_VIRTUAL_WARN("set_tower_type()"); return  ; }
 
   
  protected:
